Fix "bin not installed" error on "composer install no-dev"
We introduced the Composer Bin package as a dev dependency, and this
required a "composer bin install" post-install/update Composer command.
However when doing "composer install no-dev" this throws a
"Command 'bin' is not defined" error.

The solution is to add a dummy "bin" command in composer.json that
will run in lieu of the Composer Bin one when doing a no-dev install.