Update `lostpassword_post` action and add missing parameter
`lostpassword_post` is a WP action that is duplicated in WooCommerce to replicate the functionality to retrieve user password in the 'My Account' pages. WP recently added the variable `$user_data` as a second parameter to this action (https://github.com/WordPress/WordPress/commit/a6cecef42f22ab635684052e0d3011e5d985e1d3). This commit simply copies this change to the version of the action that we maintain. Similar to what was done in 1a99235dc85193b43b43e03fa225aef9a69fde89 when a first parameter as added to the action in WP and we had to do the same in WC core.